Added "previous card" button (no history, just reverse order)
"Next card" and "previous card" are not the right words. More like "take from top of deck" and "take from bottom of deck" -- the deck is a cycle, and you can traverse in either order. Probably there should be an actual *history*, but idk how to make it work with filters changing & such.